+++ quam/core/quam_classes.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from collections.abc import Iterable
 import sys
 import warnings
@@ -225,8 +226,6 @@ class QuamBase(ReferenceClass):
     args:
         parent: The parent of this object. This is automatically set when adding
             this object to another QuamBase object.
-        _root: The QuamRoot object. This is automatically set when instantiating
-            a QuamRoot object.
         config_settings: A dictionary of configuration settings for this object.
             This is used by [`QuamRoot.generate_config`][quam.core.quam_classes.QuamRoot.generate_config]
             to determine the order in which to add the components to the QUA config.
@@ -238,8 +237,7 @@ class QuamBase(ReferenceClass):
     """
 
     parent: ClassVar["QuamBase"] = ParentDescriptor()
-    _root: ClassVar["QuamRoot"] = None
-
+    _last_instantiated_root: ClassVar[Optional["QuamRoot"]] = None
     config_settings: ClassVar[Dict[str, Any]] = None
 
     def __init__(self):
@@ -268,6 +266,28 @@ class QuamBase(ReferenceClass):
         """
         assert is_dataclass(self)
         return [data_field.name for data_field in fields(self)]
+
+    def get_root(self) -> Optional[QuamRoot]:
+        """Get the QuamRoot object of this object.
+
+        This function recursively searches the parent chain for a QuamRoot object.
+        If no QuamRoot object is found, it will return the last instantiated QuamRoot
+        if it exists, else None.
+
+        Returns:
+            The root of this object, or None if no root is found.
+        """
+        if self.parent is not None:
+            return self.parent.get_root()
+
+        if self._last_instantiated_root is not None:
+            warnings.warn(
+                f"This component is not part of any QuamRoot, using last "
+                f"instantiated QuamRoot. This is not recommended as it may lead to "
+                f"unexpected behaviour. Component: {self.__class__.__name__}"
+            )
+            return self._last_instantiated_root
+        return None
 
     def get_attr_name(self, attr_val: Any) -> str:
         """Get the name of an attribute that matches the value.
@@ -539,7 +559,10 @@ class QuamBase(ReferenceClass):
         if not string_reference.is_reference(reference):
             return reference
 
-        if string_reference.is_absolute_reference(reference) and self._root is None:
+        if (
+            string_reference.is_absolute_reference(reference)
+            and self.get_root() is None
+        ):
             warnings.warn(
                 f"No QuamRoot initialized, cannot retrieve reference {reference}"
                 f" from {self.__class__.__name__}"
@@ -548,7 +571,7 @@ class QuamBase(ReferenceClass):
 
         try:
             return string_reference.get_referenced_value(
-                self, reference, root=self._root
+                self, reference, root=self.get_root()
             )
         except ValueError as e:
             try:
@@ -564,7 +587,7 @@ class QuamBase(ReferenceClass):
         Args:
             indent: The number of spaces to indent the summary.
         """
-        if self._root is self:
+        if self.get_root() is self:
             full_name = "QuAM:"
         elif self.parent is None:
             full_name = f"{self.__class__.__name__} (parent unknown):"
@@ -651,22 +674,19 @@ class QuamRoot(QuamBase):
         This class should not be used directly, but should generally be subclassed and
         made a dataclass. The dataclass fields should correspond to the QuAM root
         structure.
-
-    Note:
-        Upon instantiating a `QuamRoot` object, it sets the class attribute
-        `QuamBase._root` to itself. This is used such that any references with an
-        absolute path are resolved from the root.
     """
 
     serialiser: AbstractSerialiser = JSONSerialiser
 
     def __post_init__(self):
-        if QuamBase._root is not None:
+        if QuamBase._last_instantiated_root is not None:
             warnings.warn(
-                "A new QuamRoot instance has been created while a previous one exists. "
-                "The previous instance should no longer be used."
+                "Multiple QuamRoot objects were instantiated. Any QuAM component will be "
+                "attached to its specific QuamRoot object. Mixing QuAM components belonging "
+                "to different QuamRoot objects is not recommended as it may lead to "
+                "unexpected results."
             )
-        QuamBase._root = self
+        QuamBase._last_instantiated_root = self
         super().__post_init__()
 
     def __setattr__(self, name, value):
@@ -686,6 +706,17 @@ class QuamRoot(QuamBase):
         if relative_path is not None:
             reference = string_reference.join_references(reference, relative_path)
         return reference
+
+    def get_root(self: QuamRootType) -> QuamRootType:
+        """Get the QuamRoot object of this object, i.e. the object itself.
+
+        This QuamRoot function overrides the QuamBase function to return the object
+        itself, rather than following the parent chain.
+
+        Returns:
+            The current QuamRoot object (self).
+        """
+        return self
 
     def save(
         self,
@@ -1057,6 +1088,8 @@ class QuamList(UserList, QuamBase):
     def __getitem__(self, i):
         elem = super().__getitem__(i)
         if isinstance(i, slice):
+            if isinstance(elem, QuamList):
+                elem.parent = self.parent
             # This automatically gets the referenced values
             return list(elem)