+++ tests/examples/superconducting_qubits/test_transmon.py
@@ -281,7 +281,7 @@ def test_instantiation_single_element():
     assert quam.qubit.id == 0
     assert quam.qubit.xy is None
 
-    assert quam.qubit._root is quam
+    assert quam.qubit.get_root() is quam
 
 
 def test_instantiation_single_nested_element():
@@ -290,6 +290,6 @@ def test_instantiation_single_nested_element():
     assert quam.qubit.xy.mixer.name == "q0.xy.mixer"
     assert quam.qubit.xy.mixer.local_oscillator_frequency == 6e9
 
-    assert quam.qubit._root is quam
-    assert quam.qubit.xy._root is quam
-    assert quam.qubit.xy.mixer._root is quam
+    assert quam.qubit.get_root() is quam
+    assert quam.qubit.xy.get_root() is quam
+    assert quam.qubit.xy.mixer.get_root() is quam